refactor: Extract Java build tool installers
GradleInstaller and MavenInstaller are big and getting bigger. Move them
into their own files.
